+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Skip validation for `service-priority` label.
+
 ## [4.0.2] - 2022-04-04
 
 ### Fixed